+++ app/src/main/AndroidManifest.xml
@@ -54,7 +54,7 @@
         </activity>
 
         <activity
-            android:name=".routes.activity.CameraV2Activity"
+            android:name=".routes.activity.CameraActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.App.Dark">
