+++ app/src/main/res/layout/activity_camera.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/clMain"
-    tools:context=".routes.activity.CameraV2Activity">
+    tools:context=".routes.activity.CameraActivity">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/viewFinder"
